kgsl: gmu: Dynamically read context record size from ucode
Currently, discrepancies can occur between the preemption context
record size in the ucode, GMU and KGSL. Instead of hard coding the
size in GMU and KGSL, dynamically parse context record sizes from
the ucode binary. Microcode binary contains a table with context
record sizes for a list of GPUs, and KGSL needs to parse it to
determine the appropriate size for a particular GPU. Use VRB to
pass the size to GMU.

Also, retain the legacy method to maintain backward compatibility.

Change-Id: I84c9dd6ba526100c74c32d77412987dc7e26aadf
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>