kgsl: adreno: Defer preemption record gmem allocation
Only allocate non-gmem portion of the preemption records at
the time of boot, when handling the mem alloc request from GMU.
Defer the allocation of gmem portion of the preemption record when
the first context is created for that ringbuffer priority level.
We need to ensure that these buffers are virtually contiguous since
GMU assumes that about these buffers. A failure to allocate the
gmem portion will be logged and will result in context creation
failure. Add ADRENO_DEFER_GMEM_ALLOC feature bit for this feature.

Change-Id: I81819f56b87310f050394bf81a054f4dadf9c3c3
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>