kgsl: hwsched: Set ADRENO_CONTEXT_FAULT flag upon GPU fault
Currently, we do not check and set ADRENO_CONTEXT_FAULT after
queuing the command, which results in IOCTL_KGSL_GPU_COMMAND
not returning the appropriate error to the user.

Add a check to ensure that if there is a gpu fault,
it returns EPROTO error to userspace. This error indicates
to userspace to start pulling the snapshot from the sysfs
node (kgsl-3d0/snapshot) into the specified path.

Change-Id: I75ac9eec87abd64e256dd1bc6e240eea04ba25f1
Signed-off-by: Sanjay Yadav <quic_sanjyada@quicinc.com>