kgsl: hwfence: Update GMU context queue header out_fence_ts
If after reset-recovery, kgsl sends any fence with SKIP_MEMSTORE
flag set, then GMU cannot update the out_fence_ts variable for
such fence. This is because the GMU cannot access the context queue
(and its header) as the context is not registered with GMU. When
this fence entry is eventually destroyed, it may appear as
unsignaled. Hence, when setting the SKIP_MEMSTORE flag for a fence,
also update the out_fence_ts from kgsl itself after when the fence
has been ack'd by the GMU.
Also, change the loop to list_for_each_entry() since the entries
are not deleted in this loop.

Change-Id: I1427b7e671cdd617c187023d0b7fe4c06ebc7d1d
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>