kgsl: hwfence: Do not increment syncobj ts during reset-replay
The syncobj timestamp is getting incremented everytime a syncobj
gets resubmitted during replay after reset. This can lead to
inconsistencies. Also, the logic of retiring a syncobj during replay
is checking if the syncobj's sw dma fences are retired. This
too can lead to inconsistencies. So, unify the logic of retiring
a syncobj through normal retire path as well as replay path.

Change-Id: I8cdeebcf7a2360e0ea966ec37cbaa8f86a1ffeb5
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>