kgsl: hwsched: Clear HFI buffers properly during reset
Currently, HFI buffers are cleared before suspending the GMU and
halting GBIF traffic, which can result in data being written to the
buffer after it has been set to zero. This leads to an inconsistent
state and asserts in the GMU firmware during recovery. Clear the HFI
buffer properly after suspending GMU and GBIF to prevent this race
condition.

Change-Id: Ic671f646a1305204b00f25a84ad8c7f1d4e4e951
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>