kgsl: hwfence: Fix hw syncobj destroy race
If GMU has retired a sync object, it will get destroyed by the hwsched
dispatcher. But, it is oblivious to whether dma fence callbacks have
also been executed for this sync object. And that creates a race where
the callbacks may execute during (or after) sync object destruction.
To fix this race, make sure we remove the callbacks before we destroy
the sync object.
Also, the logic to see whether a sync object is to be replayed (after
hard reset) is inverted, which is causing un-retired sync objects to
get destroyed while retired sync objects are being re-sent to GMU again.
This can cause "Missing sync object" error signature. Fix the logic
to only send un-retired sync objects to GMU after hard reset.
Also add a check to not re-send sync objects that belong to detached or
invalidated contexts.

Change-Id: If02b01ff5d4115c50817122f221a486d83909ac5
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>