kgsl: hwfence: Always allow kgsl input fences to be sent to GMU
If we receive an input fence which is a kgsl fence, no matter if
it is a hw fence, allow it to be sent to GMU as GMU has access
to memstore to be able to figure out if the fence is retired.
This is especially helpful in use-cases which create duplicate
or out-of-order kgsl fences to coordinate between different
kgsl contexts.

Change-Id: I30f1c4e89fd3f3608545aa347ae17800cd8b4f21
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>