kgsl: Support firmware-name DT property
Update zap shader loading logic to try the 'firmware-name' DT property
first and if it is missing, fallback to zap_name entry from the gpulist.
If both sources are missing, assume that the platform allows programing
RBBM_SECVID_TRUST_CNTL to switch secure state.

For all new GPUs, the zap_name entry in the GPU list is deprecated, while
existing entries are retained for backward compatibility.

Change-Id: I332e234cf0a47621da3de29ad07bbb586a6ca9c9
Signed-off-by: Kaushal Sanadhya <ksanadhy@qti.qualcomm.com>