kgsl: hwfence: Do not wait for hw fence timer to finish
The hw fence timer gets deleted when we hit a fault. Sometimes
this function may be invoked in interrupt context, where using
the sync version of timer delete APIs throws a warning. Since
we don't require the timer to complete (if it was running
concurrently) before we handle a fault, use the regular timer
delete API.

Change-Id: I3828599158cd0158617fdacbdd7f8afd97bb4e10
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>