kgsl: Handle errors when creating timeline AUX commands
During memory pressure or other cases, it is possible for the
kgsl_drawobj_add_timeline() to fail to register its events. The references
to the timelines are removed but the count is not reset. This leads to
errors in the subsequent cleanup. Reset the count of timelines to 0 so that
the structure can be safely cleaned up.

Change-Id: Ic32e8926bd2be71df0c12d0347f5f1f89d24ca40
Signed-off-by: Scott Bauer <sbauer@qti.qualcomm.com>
Signed-off-by: Lynus Vaz <lvaz@qti.qualcomm.com>