kgsl: reclaim: Don't update the memdesc's vmas when migrating
The vmas in the memdesc represent the mappings of the memdesc in the mm
space. Since they are already inserted in the address space, do not modify
the vma backing file. Since migrating pages also requires us to unmap
pages, take the mmap write lock to lock out other accesses.

Change-Id: Ib46e825651fc2d0ae4f4d40fd37ce4130be74a20
Signed-off-by: Shiv Kumar <shikum@qti.qualcomm.com>
Signed-off-by: Lynus Vaz <lvaz@qti.qualcomm.com>