kgsl: Simplify the locking for memdesc lock
The migrate operation already takes the exclusive mmap lock, thus
preventing concurrent vm_fault operations. Therefore, it is not necessary
to hold the memdesc lock for the same operations. Reduce the amount of time
we hold the memdesc lock since it is safe to do so. This also allows us to
use a spinlock instead of an rwsem.

Change-Id: I35e453cd8c64883eb17b4c71cf1bb7426cc2c763
Signed-off-by: Lynus Vaz <lvaz@qti.qualcomm.com>
(cherry picked from commit d785c6e92badd81552e5500f571d9702bb5f591f)