kgsl: Add back support for standard compatible strings
Support for standard compatible strings was removed in commit
6c80bec6478b ("kgsl: Remove support for standard compatible strings").
Add back support for them.  When nodes with both standard
and legacy compatible strings are available, probe the node with
legacy compatible string and return error for the node with standard
compatible string. Also when both the nodes are present and are tried
for probe, the iommu group may contain more than one device. In this
case iommu_attach_device() fails. Hence, use iommu_attach_group().

Change-Id: I3549e0ada7567a57861a20cd214c818ed2054630
Signed-off-by: Puranam V G Tejaswi <pvgtejas@qti.qualcomm.com>