kgsl: hwfence: Destroy inflight hardware fences in context detach path
There are cases where a context gets detached and it was not registered
with GMU or GMU is powered down. In some such corner cases, when this
happens alongwith a gpu fault reset/recovery sequence, there may be
inflight hardware fences from this context which are not destroyed.
This context never gets destroyed, as each hardware fence entry takes a
refcount on the context. To address this, make sure to destroy the
hardware fence entries in this path. Add a warning if either of the
per context hw fence lists are not empty.

Change-Id: Id6fd6eb3e4ef7f2e16027bc3d850bba11d624f28
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>
(cherry picked from commit d6fb5ddd57ba7ba07ff2b806ff741aae8714c97c)