kgsl: Ensure IRQ name string persists using devm_kasprintf
Use devm_kasprintf() to allocate the IRQ name string when requesting
interrupt based on IRQ index. This ensures the string remains valid for
the lifetime of the device and avoids issues with temporary stack-based
buffers since the IRQ name string is passed to devm_request_irq API and
may be accessed later.

Change-Id: Iaa8d1f8d734a364ae8e94ed72faf5301b26770f8
Signed-off-by: Kaushal Sanadhya <ksanadhy@qti.qualcomm.com>