kgsl: hwsched: Abort GPU suspend if context unregister is not received
There is a potential race condition during pm_suspend where the context
unregister event may not be received during GPU suspend. If this occurs,
the active_cnt remains at 1 after transitioning to the suspend state.
When the hwsched worker thread attempts to submit a workload, it skips
the GPU bootup check because active_cnt is not zero, and tries to submit
the workload while the GPU is in slumber, leading to unwanted behavior.

This change aborts the GPU suspend if active_cnt is greater than zero,
preventing submission of workloads in the slumber state.

Change-Id: I581f9998a7f3b8a32e752374c5359a82128bd25c
Signed-off-by: Shiv Kumar <shikum@qti.qualcomm.com>