kgsl: Apply qos constraints after first boot
Holding device mutex to apply qos constraint early boot can cause mutex
deadlock. To prevent this, active count was used for an early return.
However, active count can be zero and the gpu still be up if the
submissions are sparse, which would cause some qos constraints to be
dropped. Hence, add a flag to mark the completion of first boot and
apply all qos constraints only after first boot is done.

Change-Id: I1696573b37b51dbe69fe44a83abf5410e1aae8fe
Signed-off-by: Mohammed Mirza Mandayappurath Manzoor Mopan <mmandaya@qti.qualcomm.com>