kgsl: hwfence: Take kgsl hw fence context refcount only once
When a hardware fence (which belongs to a kgsl context say ctx_a) is
being waited on by another kgsl context say ctx_b, and say a
reset-recovery happens, we end up taking an extra refcount on ctx_a
when resubmitting the same hardware fence object back to the context
queue again. To address this, make sure we only take the context
refcount when sending it to the context queue for the very first
time.

Change-Id: Icef101f84865d31b4ac444a0533080c44e603c1f
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>
Signed-off-by: Kaushik Yalla <kyalla@qti.qualcomm.com>