kgsl: hwsched: Abort GPU suspend if context unregister is not received
There is a potential race condition during pm_suspend where the context
unregister event may not be received during GPU suspend. If this occurs,
the active_cnt remains at 1 after transitioning to the suspend state.
When the hwsched worker thread attempts to submit a workload, it skips
the GPU bootup check because active_cnt is not zero, and tries to submit
the workload while the GPU is in slumber, leading to unwanted behavior.

This change aborts the GPU suspend if active_cnt is greater than zero,
preventing submission of workloads in the slumber state.

Change-Id: Ib1373d377792a76732ac2569547d6f8cf28bd85b
Signed-off-by: Kaushik Yalla <kyalla@qti.qualcomm.com>