kgsl: gen8: Add flag for non RBBM counters and to reserve GMU counters
GMU can maintain refcounts for perfcounters and request allocation
or deallocation to KMD based on refcounts. Add a separate flag
to track the request from GMU and release the counter only when nobody
has the claim on the counter.

There are non RBBM counters which need not be restored by CP. Add a
separate flag to mark these counters to prevent being added to
dynamic list.

Change-Id: Ibe63de600bbadaace5a22d03da7e97c547962ddf
Signed-off-by: Mohammed Mirza Mandayappurath Manzoor Mopan <mmandaya@qti.qualcomm.com>