kgsl: Fix uninitialized val passed to kgsl_regmap_read_poll_timeout()
The read_poll_timeout() macro and its wrappers, such as
kgsl_regmap_read_poll_timeout(), check the polling condition
before the first read of the register. If the variable used in
the condition is not initialized, it can contain a garbage value
that coincidentally satisfies the condition. This leads to the
poll succeeding prematurely without ensuring the hardware is in
the correct state.

Change-Id: I2253d289e10fecfeaaafe341ffdec350f494b268
Signed-off-by: Chandra Vamsi Yekkaluri <cyekkalu@qti.qualcomm.com>
Signed-off-by: Jayachandra Reddy Sura <jsura@qti.qualcomm.com>