kgsl: gen8: Support encrypted debugbus data on gen8 targets
On fused devices there is no direct way of capturing essential debugbus
data. The Qualcomm Enclave Control Processor (QECP) is a new HW which
can encrypt and capture debugbus on fused devices. To achieve this we
use the QMI interface. QMI messages are used to communicate between
KGSL and QECP. Add capability in KGSL to communicate with QECP through
the QMI interface and capture the debugbus on fused devices.

Change-Id: I1db328fa3b37a76a9a0816d3ab3ddbf99fb59ac4
Signed-off-by: Urvashi Agrawal <urvaagra@qti.qualcomm.com>
Signed-off-by: Lynus Vaz <lvaz@qti.qualcomm.com>
Signed-off-by: David Werner <daviwern@qti.qualcomm.com>