kgsl: gen8: Add support for KGSL_CMDBATCH_USES_MALU flag
When userspace sends a submission to kgsl that uses MALU, it
will set KGSL_CMDBATCH_USES_MALU flag. Make sure that gen8
target has malu and that the GMU FW also supports MALU.
If so, plumb this flag over to the GMU FW which can then insert
POWER_CONTROL PM4 packet to the cmdstream to power up MALU.
If the GMU FW doesn't support MALU, then log a message during boot.
As an extra caution, log an error (once) and return error to userspace
if userspace sets this flag but the GMU FW doesn't support
MALU. Add this support for swsched as well.

Change-Id: I7d218ed362cfbc78b1f8c0468710bbe0a38d7e15
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>