kgsl: gen8: Add support for KGSL_PROP_IS_MALU_ENABLED
Targets that have a physical MALU may have it fused out. So
add support for querying the fuse register to see if MALU is
fused out or not. Add support for enabling sw fuse violation
interrupts for MALU. Add the property so that userspace can
query whether MALU is fused out or not. Also move the placement
of POWER_CONTROL packet in swsched to avoid opcode errors.

Change-Id: I7b500c08e81f2e0a82112f9c4f127fad5f834eb5
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>