kgsl: gen8: Trigger MX voltage change during first boot
The VRM IDs for the graphics rails has to be programmed
before any of these gpu voltages are turned on. For targets
that have the GMU on MX rail, the VRM programming happens after
MX GDSC has been turned on (because these registers are on
the MX rail). Because of this, if GMU FW ends up voting for the
same MX corner during voltage calibration, there will be no
voltage broadcast to GMU for MX rail via the broadcast register.
This would cause GMU boot calibration to fail. To address this,
bump up the GMU frequency to the highest corner in order to
affect a change in MX voltage, which would then populate the
MX voltage in the broadcast register, allowing GMU FW calibration
to not fail. However, if for some reason this fails to populate
the broadcast register, then log an error and return error.

Change-Id: I1a65bd88a475c6be627b8d3d78fca4db1017d2f4
Signed-off-by: Harshdeep Dhatt <hdhatt@qti.qualcomm.com>