kgsl: Decouple iommu_unmap() from delaying kgsl_open()
During process private cleanup, iommu_unmap() in gmu_core_free_block()
can take a while to complete. If there is a concurrent thread in the
same process which is calling kgsl_open(), kgsl will return -EEXIST
if the process private is not cleaned up in given time. So, make sure
we remove the process private structure from the process private list
before calling gmu_core_free_block(). Move gmu_core_free_block() to the
very end of process private clean up so that it doesn't delay the clean
up of rest of the process private resources.

Change-Id: Id29bccd0fbaef6c46f759242d0a2a2507135384a
Signed-off-by: Shiv Kumar <shikum@qti.qualcomm.com>
Signed-off-by: Nagababu Pamarthi <npamarth@qti.qualcomm.com>