kgsl: gen8: Enable deferred GMEM allocation for gen8_11_0
Enable the ADRENO_DEFER_GMEM_ALLOC feature for the gen8_11_0 target.
This feature delays the allocation of the GMEM portion of each
preemption record until first use. That is, until the first context
gets created for a specific ring buffer priority level. In doing so,
it avoids performing this significant allocation at boot time.

Change-Id: I0a77501c20d18412f72dee2cb8ba4bdb2c8b4b61
Signed-off-by: David Werner <daviwern@qti.qualcomm.com>