kgsl: Prevent sign extension on alignments
Currently we allow alignments from userland which
eventually can be sign extended which cause havoc
down in the memory management routines. Fix the
issue by changing the return type of
kgsl_memdesc_get_align in case we use it differently
in the future and expliticly start with a unsigned
type for bit shifts.

Change-Id: I03ad7e260434a5b5a3ea08c006128bc59cbcd617
Signed-off-by: Scott Bauer <sbauer@qti.qualcomm.com>